cog_search_API_2/cog_search_API_zensar_venture_PoC/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2.4
django-cors-headers==3.11.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.1
openpyxl==3.0.9
pandas==1.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
xlrd==2.0.1
